+++ dalesbred/src/main/java/fi/evident/dalesbred/support/spring/DalesbredConfigurationSupport.java
@@ -1,0 +1,61 @@
+/*
+ * Copyright (c) 2013 Evident Solutions Oy
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package fi.evident.dalesbred.support.spring;
+
+import fi.evident.dalesbred.Database;
+import fi.evident.dalesbred.dialects.Dialect;
+import org.jetbrains.annotations.Nullable;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.transaction.PlatformTransactionManager;
+
+import javax.sql.DataSource;
+
+/**
+ * Abstract base class for Spring Configuration to provide Dalesbred integration with Spring services.
+ */
+public abstract class DalesbredConfigurationSupport {
+
+    @Autowired
+    private DataSource dataSource;
+
+    @Autowired
+    private PlatformTransactionManager platformTransactionManager;
+
+    @Bean
+    public Database dalesbredDatabase() {
+        Dialect dialect = dialect();
+        if (dialect == null)
+            dialect = Dialect.detect(dataSource);
+
+        return new Database(new SpringTransactionManager(dataSource, platformTransactionManager), dialect);
+    }
+
+    /**
+     * Subclasses can override this to return the {@link Dialect} to use. By default
+     * {@code null} is returned, which means that dialect is auto-detected.
+     */
+    @Nullable
+    protected Dialect dialect() {
+        return null;
+    }
+}